refactor(validation): modifica validação para existsById

+++ src/main/java/com/delegrego/exemplo_jpa_2/service/FuncionarioService.java
@@ -127,7 +127,9 @@ public class FuncionarioService {
 	 */
 	public void deletarFuncionario(int id) {
 
-		funcionarioRepo.findById(id).orElseThrow(() -> new RuntimeException("Funcionário não existe"));
+		if (!funcionarioRepo.existsById(id)) {
+			throw new RuntimeException("Funcionário não existe");
+		}
 
 		funcionarioRepo.deleteById(id);
 	}
